Added fig.map_overlay_unflattened_lines
The standard line function includes the conversion from the dynlib
line storage format to lines as they can be plotted. Some line data
is already directly usable for plotting, so this variant of the
existing function bypasses the conversion step.

Documentation is adapted to clarify the two variants of the function.