Slight refactor to make q_special generally available
q_special is now a kwarg used generally in get_normalized_from_file, and
is thus available in all data request functions relying on this
function. This includes in particuar get_composite, where this function
is now available for the first time. For other request functions, this
refactor simplifies the code somewhat and avoids code duplication.