Added new cyclone detection and tracking algorithm to dynlib
The algorithm is a slightly adapted re-implementation of the Melbourne
algorithm. As in the Melbourne algorithm, cyclones are detected at the
location of a minima in the Laplacian of MSLP. Both open and closed
systems are detected.

In contrast to the orginal Melbourne scheme, tracks are smoothed by a
Kalman smoother which should yield numerically less noisy estimates of
propagation direction and speed.